src/hdmi_control/requirements.txt
click==6.4
pyserial==3.4
